Challenge3/aaoifi-enhancement-system/requirements.txt
# AAOIFI Standards Enhancement System Requirements
# Core Dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.23
langchain-openai==0.0.5
langchain-google-genai==0.0.6
openai==1.78.0
google-api-python-client==2.108.0
google-generativeai==0.3.2

# Document Processing
PyPDF2==3.0.1
pypdf==5.4.0
tiktoken==0.5.2

# Data Processing
numpy==1.26.4
pandas==1.3.0
scikit-learn==0.24.2

# Visualization Dependencies
matplotlib==3.7.2
networkx==3.1

# Database & Vector Storage
chromadb==0.4.22
faiss-cpu==1.11.0
SQLAlchemy==2.0.40

# Environment & Configuration
python-dotenv==1.1.0
pydantic==2.11.4

# HTTP & API Clients
requests==2.32.3
httpx==0.28.1

# Utilities
tqdm==4.67.1
tenacity==8.5.0
PyYAML==6.0.2
dataclasses-json==0.6.7
typing-inspect==0.9.0
termcolor==3.1.0

# Testing
pytest==6.2.4

# Optional: For Command Line Interface
typer==0.15.3
rich==14.0.0